Extract left sidebar into own component in guide articles
